+++ src/app/components/home/AssistantStatus/types.ts
@@ -1,0 +1,9 @@
+import { Message } from "../WindowChat/types";
+
+export type AssistantStatusProps = {
+  status: string;
+  windowExpand: boolean;
+  isLoading: boolean;
+  setWindowExpand: React.Dispatch<React.SetStateAction<boolean>>;
+  setChatHistory: React.Dispatch<React.SetStateAction<Message[]>>;
+};
