+++ src/app/components/home/ClearChat/ClearChat.tsx
@@ -14,7 +14,7 @@ const ClearChatHistory: React.FC<ClearChatHistoryProps> = ({
 
   return (
     <button
-      className="p-2 text-sm text-gray-600/50 my-0 mx-2 xs:mx-2 py-1 px-2 rounded-lg transition duration-300 hover:shadow-md hover:bg-gray-100/30 hover:text-gray-600/70"
+      className="text-sm text-gray-600/60 rounded-lg transition duration-300 hover:text-gray-600/90"
       onClick={clearHistory}
     >
       Clear
